module_3/requirements.txt
flask>=3.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
beautifulsoup4>=4.12.3
reportlab>=4.0.0
